Reworked the seralization model.
Now model objects don't use Codable for seralization. Instead the model objects
are responsible for generating JSON to go to ES as well as initalizing an
object with JSON from ES. This may sound lame and in a way it is. But it's the
best solution that I can come up with at the time that satisfies the
requirements from all sides.

Also including some initial work on ES index configuration. Not sure how much
of this will persist.