Reworked pretty much every aspect of how analyzers, tokenizers, character filters and normalizers.
The goal here is to remove the need to explicitly define stuff in the analysis
section and then use the string name in the property settings. Instead,
properties are defined and the normalizer/analyzer definition is passed in
directly. With a touch of hard work those definitions are then automatically
configured in the analysis section.

Still need to implement token filters.